DBZ-7066 Code reorganization
