+++ debezium-connector-mongodb/src/test/java/io/debezium/connector/mongodb/transforms/ExtractNewDocumentStateTestIT.java
@@ -41,24 +41,21 @@ import io.debezium.config.Configuration;
 import io.debezium.connector.mongodb.Module;
 import io.debezium.connector.mongodb.MongoDbConnectorConfig;
 import io.debezium.data.Envelope;
-import io.debezium.data.Envelope.Operation;
-import io.debezium.data.SchemaUtil;
 import io.debezium.data.VerifyRecord;
 import io.debezium.doc.FixFor;
 import io.debezium.junit.SkipWhenDatabaseVersion;
 import io.debezium.transforms.ExtractNewRecordStateConfigDefinition;
+import io.debezium.transforms.extractnewstate.DefaultDeleteHandlingStrategy;
 import io.debezium.util.Collect;
 
 /**
- * Integration test for {@link ExtractNewDocumentState}. It sends operations into
+ * Integration test for {@link ExtractNewDocumentState} and {@link DefaultDeleteHandlingStrategy}. It sends operations into
  * MongoDB and listens on messages that are generated by Debezium plug-in. The messages
  * are then run through the SMT itself.
  *
- * @author Jiri Pechanec
+ * @author Harvey Yue
  */
 public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentStateTestIT {
-
-    // todo: Verify that shard logic added in c84a8a28b472b7c169a8b265905a64db0dbb69e3 doesn't apply now.
 
     @Override
     protected String getCollectionName() {
@@ -68,6 +65,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @Test
     @FixFor("DBZ-563")
     public void shouldDropTombstoneByDefault() throws InterruptedException {
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "drop"));
         // First insert
         try (var client = connect()) {
             client.getDatabase(DB_NAME).getCollection(this.getCollectionName())
@@ -84,7 +82,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         }
 
         // First delete record to arrive is coming from the oplog
-        SourceRecord firstRecord = getRecordByOperation(Operation.DELETE);
+        SourceRecord firstRecord = getRecordByOperation(Envelope.Operation.DELETE);
         final SourceRecord transformedDelete = transformation.apply(firstRecord);
         assertThat(transformedDelete).isNull();
 
@@ -100,8 +98,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @Test
     public void shouldTransformEvents() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
-        transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
-        transformationConfig.put(HANDLE_DELETES, "none");
+        // transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
+        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -241,17 +239,19 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final SourceRecord tombstoneRecord = records.recordsForTopic(this.topicName()).get(1);
         final SourceRecord transformedTombstone = transformation.apply(tombstoneRecord);
 
-        assertThat(transformedTombstone.value()).isNull();
+        // assertThat(transformedTombstone.value()).isNull();
+        assertThat(transformedTombstone).isNull();
 
         // Assert deletion preserves key
-        assertThat(SchemaUtil.asString(transformedDelete.keySchema())).isEqualTo(SchemaUtil.asString(transformedTombstone.keySchema()));
-        assertThat(transformedDelete.key().toString()).isEqualTo(transformedTombstone.key().toString());
+        // assertThat(SchemaUtil.asString(transformedDelete.keySchema())).isEqualTo(SchemaUtil.asString(transformedTombstone.keySchema()));
+        // assertThat(transformedDelete.key().toString()).isEqualTo(transformedTombstone.key().toString());
     }
 
     @Test
     @FixFor("DBZ-1767")
     public void shouldSupportDbRef() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
+        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformationConfig.put("array.encoding", "array");
         transformationConfig.put("operation.header", "true");
         transformationConfig.put("field.name.adjustment.mode", "avro");
@@ -278,6 +278,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @FixFor("DBZ-2680")
     public void shouldSupportSubSanitizeFieldName() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
+        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformationConfig.put("array.encoding", "array");
         transformationConfig.put("operation.header", "true");
         transformationConfig.put("field.name.adjustment.mode", "avro");
@@ -313,6 +314,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         props.put(ADD_FIELDS, "ord , db,rs");
         transformation.configure(props);
 
@@ -361,7 +363,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "ord,db,rs");
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         transformation.configure(props);
 
         // insert
@@ -406,6 +408,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         props.put(ADD_HEADERS, "op");
         transformation.configure(props);
 
@@ -435,7 +438,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         // then assert operation header is insert
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.CREATE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.CREATE.code());
 
         // acquire key and value Structs
         Struct key = (Struct) transformed.key();
@@ -466,6 +469,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @Test
     public void shouldTransformRecordForInsertEventWithComplexIdType() throws InterruptedException {
         waitForStreamingRunning();
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         Document obj = new Document()
                 .append("_id", new Document()
@@ -516,6 +520,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         props.put(ADD_HEADERS, "op");
         transformation.configure(props);
 
@@ -555,7 +560,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         // then assert operation header is update
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.UPDATE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.UPDATE.code());
 
         // acquire key and value Structs
         Struct key = (Struct) transformed.key();
@@ -584,6 +589,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
                 .build();
         restartConnectorWithConfig(config);
         waitForStreamingRunning();
+
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -647,6 +654,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         restartConnectorWithConfig(config);
         waitForStreamingRunning();
 
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
+
         ObjectId objId = new ObjectId();
         Document obj = new Document()
                 .append("_id", objId)
@@ -699,6 +708,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     public void shouldGenerateRecordForUpdateEventWithUnset() throws InterruptedException {
         waitForStreamingRunning();
 
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
+
         ObjectId objId = new ObjectId();
         Document obj = new Document()
                 .append("_id", objId)
@@ -749,6 +760,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @FixFor("DBZ-612")
     public void shouldGenerateRecordForUnsetOnlyUpdateEvent() throws InterruptedException {
         waitForStreamingRunning();
+
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -803,7 +816,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "none");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -854,7 +867,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op");
-        props.put(DROP_TOMBSTONE, "false");
+        // props.put(DROP_TOMBSTONE, "false");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -895,7 +909,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         // then assert operation header is delete
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.DELETE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.DELETE.code());
 
         assertThat(value).isNull();
     }
@@ -905,6 +919,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     public void shouldDropDeleteMessagesByDefault() throws InterruptedException {
         restartConnectorWithoutEmittingTombstones();
         waitForStreamingRunning();
+
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "drop"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -943,7 +959,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -990,7 +1006,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1035,7 +1051,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op");
-        props.put(HANDLE_DELETES, "none");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1067,7 +1083,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         // then assert operation header is delete
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.DELETE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.DELETE.code());
 
         // acquire key and value Structs
         Struct key = (Struct) transformed.key();
@@ -1085,6 +1101,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @FixFor("DBZ-971")
     public void shouldPropagatePreviousRecordHeaders() throws InterruptedException {
         waitForStreamingRunning();
+
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "drop"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -1127,6 +1145,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     @Test
     public void shouldNotFlattenTransformRecordForInsertEvent() throws InterruptedException {
         waitForStreamingRunning();
+
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -1181,6 +1201,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(FLATTEN_STRUCT, "true");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1232,6 +1253,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> props = new HashMap<>();
         props.put(FLATTEN_STRUCT, "true");
         props.put(DELIMITER, "-");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1283,6 +1305,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> props = new HashMap<>();
         props.put(FLATTEN_STRUCT, "true");
         props.put(DELIMITER, "-");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1351,12 +1374,13 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
         assertThat(transformed.headers()).hasSize(1);
-        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Envelope.Operation.CREATE.code());
     }
 
     @Test
@@ -1366,12 +1390,13 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op,id");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
         assertThat(transformed.headers()).hasSize(2);
-        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Envelope.Operation.CREATE.code());
         assertThat(getSourceRecordHeaderByKey(transformed, "__id")).isNull();
     }
 
@@ -1383,12 +1408,13 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op,source.rs,source.collection");
         props.put(ADD_HEADERS_PREFIX, "prefix.");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
         assertThat(transformed.headers()).hasSize(3);
-        assertThat(getSourceRecordHeaderByKey(transformed, "prefix.op")).isEqualTo(Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, "prefix.op")).isEqualTo(Envelope.Operation.CREATE.code());
         assertThat(getSourceRecordHeaderByKey(transformed, "prefix.source_rs")).isEqualTo("rs0");
         assertThat(getSourceRecordHeaderByKey(transformed, "prefix.source_collection")).isEqualTo(getCollectionName());
     }
@@ -1400,11 +1426,12 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.CREATE.code());
     }
 
     @Test
@@ -1419,14 +1446,15 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         props.put(ADD_FIELDS_PREFIX, fieldPrefix);
         props.put(ADD_HEADERS, "op:OPERATION");
         props.put(ADD_HEADERS_PREFIX, headerPrefix);
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
 
-        assertThat(((Struct) transformed.value()).get(fieldPrefix + "OP")).isEqualTo(Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get(fieldPrefix + "OP")).isEqualTo(Envelope.Operation.CREATE.code());
         assertThat(transformed.headers()).hasSize(1);
-        assertThat(getSourceRecordHeaderByKey(transformed, headerPrefix + "OPERATION")).isEqualTo(Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, headerPrefix + "OPERATION")).isEqualTo(Envelope.Operation.CREATE.code());
     }
 
     @Test
@@ -1437,11 +1465,12 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op , ts_ms");
         props.put(ADD_FIELDS_PREFIX, "prefix.");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("prefix.op")).isEqualTo(Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("prefix.op")).isEqualTo(Envelope.Operation.CREATE.code());
         assertThat(((Struct) transformed.value()).get("prefix.ts_ms")).isNotNull();
     }
 
@@ -1452,11 +1481,12 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op,id");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.CREATE.code());
         assertThat(((Struct) transformed.value()).get("__id")).isNull();
     }
 
@@ -1467,11 +1497,12 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op,source.rs,source.collection");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.CREATE.code());
         assertThat(((Struct) transformed.value()).get("__source_rs")).isEqualTo("rs0");
         assertThat(((Struct) transformed.value()).get("__source_collection")).isEqualTo(getCollectionName());
     }
@@ -1482,14 +1513,14 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op");
         transformation.configure(props);
 
         final SourceRecord deleteRecord = createDeleteRecordWithTombstone().allRecordsInOrder().get(0);
         final SourceRecord transformed = transformation.apply(deleteRecord);
         assertThat(((Struct) transformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
     }
 
     @Test
@@ -1498,14 +1529,14 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op,ts_ms");
         transformation.configure(props);
 
         final SourceRecord deleteRecord = createDeleteRecordWithTombstone().allRecordsInOrder().get(0);
         final SourceRecord transformed = transformation.apply(deleteRecord);
         assertThat(((Struct) transformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
         assertThat(((Struct) transformed.value()).get("__ts_ms")).isNotNull();
     }
 
@@ -1515,14 +1546,14 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op,source.rs,source.collection");
         transformation.configure(props);
 
         final SourceRecord deleteRecord = createDeleteRecordWithTombstone().allRecordsInOrder().get(0);
         final SourceRecord transformed = transformation.apply(deleteRecord);
         assertThat(((Struct) transformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
         assertThat(((Struct) transformed.value()).get("__source_rs")).isEqualTo("rs0");
         assertThat(((Struct) transformed.value()).get("__source_collection")).isEqualTo(getCollectionName());
     }
@@ -1533,9 +1564,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
         props.put(ADD_FIELDS, "op,ts_ms");
-        props.put(DROP_TOMBSTONE, "false");
         transformation.configure(props);
 
         final SourceRecords records = createDeleteRecordWithTombstone();
@@ -1543,7 +1573,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final SourceRecord deleteRecord = records.allRecordsInOrder().get(0);
         final SourceRecord deleteTransformed = transformation.apply(deleteRecord);
         assertThat(((Struct) deleteTransformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) deleteTransformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
+        assertThat(((Struct) deleteTransformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
         assertThat(((Struct) deleteTransformed.value()).get("__ts_ms")).isNotNull();
 
         final SourceRecord tombstoneRecord = records.allRecordsInOrder().get(1);
@@ -1557,6 +1587,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> transformationConfig = new HashMap<>();
         transformationConfig.put("array.encoding", "array");
         transformationConfig.put("field.name.adjustment.mode", "avro");
+        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -1610,6 +1641,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "updateDescription.updatedFields");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         // Perform transformation
@@ -1653,6 +1685,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> props = new HashMap<>();
         props.put(ARRAY_ENCODING, "array");
         props.put(ADD_FIELDS, "op,source.ts_ms");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecords records = createCreateRecordFromJson("dbz-2316.json");
@@ -1666,8 +1699,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     public void testMatrixType() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
         transformationConfig.put("array.encoding", "array");
-        transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
-        transformationConfig.put(HANDLE_DELETES, "none");
+        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -1762,8 +1794,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     public void testMatrixArrayAsDocumentType() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
         transformationConfig.put("array.encoding", "document");
-        transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
-        transformationConfig.put(HANDLE_DELETES, "none");
+        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -1828,6 +1859,8 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
     public void shouldSupportNestedArrays() throws InterruptedException {
         waitForStreamingRunning();
 
+        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
+
         // Test insert
         try (var client = connect()) {
             client.getDatabase(DB_NAME).getCollection(this.getCollectionName())
@@ -1889,6 +1922,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "updateDescription.updatedFields,nonexistentField,version");
         props.put(ADD_HEADERS_PREFIX, "prefix.");
+        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1943,7 +1977,7 @@ public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentSta
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_DELETES, "rewrite");
+        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
