+++ debezium-connector-mongodb/src/test/java/io/debezium/connector/mongodb/transforms/LegacyExtractNewDocumentStateTestIT.java
@@ -41,21 +41,24 @@ import io.debezium.config.Configuration;
 import io.debezium.connector.mongodb.Module;
 import io.debezium.connector.mongodb.MongoDbConnectorConfig;
 import io.debezium.data.Envelope;
+import io.debezium.data.Envelope.Operation;
+import io.debezium.data.SchemaUtil;
 import io.debezium.data.VerifyRecord;
 import io.debezium.doc.FixFor;
 import io.debezium.junit.SkipWhenDatabaseVersion;
 import io.debezium.transforms.ExtractNewRecordStateConfigDefinition;
-import io.debezium.transforms.extractnewstate.DefaultDeleteHandlingStrategy;
 import io.debezium.util.Collect;
 
 /**
- * Integration test for {@link ExtractNewDocumentState} and {@link DefaultDeleteHandlingStrategy}. It sends operations into
+ * Integration test for {@link ExtractNewDocumentState}. It sends operations into
  * MongoDB and listens on messages that are generated by Debezium plug-in. The messages
  * are then run through the SMT itself.
  *
- * @author Harvey Yue
+ * @author Jiri Pechanec
  */
-public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentStateTestIT {
+public class LegacyExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentStateTestIT {
+
+    // todo: Verify that shard logic added in c84a8a28b472b7c169a8b265905a64db0dbb69e3 doesn't apply now.
 
     @Override
     protected String getCollectionName() {
@@ -65,7 +68,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @Test
     @FixFor("DBZ-563")
     public void shouldDropTombstoneByDefault() throws InterruptedException {
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "drop"));
         // First insert
         try (var client = connect()) {
             client.getDatabase(DB_NAME).getCollection(this.getCollectionName())
@@ -82,7 +84,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         }
 
         // First delete record to arrive is coming from the oplog
-        SourceRecord firstRecord = getRecordByOperation(Envelope.Operation.DELETE);
+        SourceRecord firstRecord = getRecordByOperation(Operation.DELETE);
         final SourceRecord transformedDelete = transformation.apply(firstRecord);
         assertThat(transformedDelete).isNull();
 
@@ -98,8 +100,8 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @Test
     public void shouldTransformEvents() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
-        // transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
-        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
+        transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
+        transformationConfig.put(HANDLE_DELETES, "none");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -239,19 +241,17 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final SourceRecord tombstoneRecord = records.recordsForTopic(this.topicName()).get(1);
         final SourceRecord transformedTombstone = transformation.apply(tombstoneRecord);
 
-        // assertThat(transformedTombstone.value()).isNull();
-        assertThat(transformedTombstone).isNull();
+        assertThat(transformedTombstone.value()).isNull();
 
         // Assert deletion preserves key
-        // assertThat(SchemaUtil.asString(transformedDelete.keySchema())).isEqualTo(SchemaUtil.asString(transformedTombstone.keySchema()));
-        // assertThat(transformedDelete.key().toString()).isEqualTo(transformedTombstone.key().toString());
+        assertThat(SchemaUtil.asString(transformedDelete.keySchema())).isEqualTo(SchemaUtil.asString(transformedTombstone.keySchema()));
+        assertThat(transformedDelete.key().toString()).isEqualTo(transformedTombstone.key().toString());
     }
 
     @Test
     @FixFor("DBZ-1767")
     public void shouldSupportDbRef() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
-        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformationConfig.put("array.encoding", "array");
         transformationConfig.put("operation.header", "true");
         transformationConfig.put("field.name.adjustment.mode", "avro");
@@ -278,7 +278,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @FixFor("DBZ-2680")
     public void shouldSupportSubSanitizeFieldName() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
-        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformationConfig.put("array.encoding", "array");
         transformationConfig.put("operation.header", "true");
         transformationConfig.put("field.name.adjustment.mode", "avro");
@@ -314,7 +313,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         props.put(ADD_FIELDS, "ord , db,rs");
         transformation.configure(props);
 
@@ -363,7 +361,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "ord,db,rs");
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         transformation.configure(props);
 
         // insert
@@ -408,7 +406,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         props.put(ADD_HEADERS, "op");
         transformation.configure(props);
 
@@ -438,7 +435,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         // then assert operation header is insert
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.CREATE.code());
 
         // acquire key and value Structs
         Struct key = (Struct) transformed.key();
@@ -469,7 +466,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @Test
     public void shouldTransformRecordForInsertEventWithComplexIdType() throws InterruptedException {
         waitForStreamingRunning();
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         Document obj = new Document()
                 .append("_id", new Document()
@@ -520,7 +516,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         props.put(ADD_HEADERS, "op");
         transformation.configure(props);
 
@@ -560,7 +555,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         // then assert operation header is update
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.UPDATE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.UPDATE.code());
 
         // acquire key and value Structs
         Struct key = (Struct) transformed.key();
@@ -589,8 +584,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
                 .build();
         restartConnectorWithConfig(config);
         waitForStreamingRunning();
-
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -654,8 +647,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         restartConnectorWithConfig(config);
         waitForStreamingRunning();
 
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
-
         ObjectId objId = new ObjectId();
         Document obj = new Document()
                 .append("_id", objId)
@@ -708,8 +699,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     public void shouldGenerateRecordForUpdateEventWithUnset() throws InterruptedException {
         waitForStreamingRunning();
 
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
-
         ObjectId objId = new ObjectId();
         Document obj = new Document()
                 .append("_id", objId)
@@ -760,8 +749,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @FixFor("DBZ-612")
     public void shouldGenerateRecordForUnsetOnlyUpdateEvent() throws InterruptedException {
         waitForStreamingRunning();
-
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -816,7 +803,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
+        props.put(HANDLE_DELETES, "none");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -867,8 +854,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op");
-        // props.put(DROP_TOMBSTONE, "false");
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
+        props.put(DROP_TOMBSTONE, "false");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -909,7 +895,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         // then assert operation header is delete
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.DELETE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.DELETE.code());
 
         assertThat(value).isNull();
     }
@@ -919,8 +905,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     public void shouldDropDeleteMessagesByDefault() throws InterruptedException {
         restartConnectorWithoutEmittingTombstones();
         waitForStreamingRunning();
-
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "drop"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -959,7 +943,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1006,7 +990,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1051,7 +1035,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
+        props.put(HANDLE_DELETES, "none");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1083,7 +1067,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         // then assert operation header is delete
         Iterator<Header> operationHeader = transformed.headers().allWithName(ExtractNewRecordStateConfigDefinition.DEBEZIUM_OPERATION_HEADER_KEY);
         assertThat((operationHeader).hasNext()).isTrue();
-        assertThat(operationHeader.next().value().toString()).isEqualTo(Envelope.Operation.DELETE.code());
+        assertThat(operationHeader.next().value().toString()).isEqualTo(Operation.DELETE.code());
 
         // acquire key and value Structs
         Struct key = (Struct) transformed.key();
@@ -1101,8 +1085,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @FixFor("DBZ-971")
     public void shouldPropagatePreviousRecordHeaders() throws InterruptedException {
         waitForStreamingRunning();
-
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "drop"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -1145,8 +1127,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     @Test
     public void shouldNotFlattenTransformRecordForInsertEvent() throws InterruptedException {
         waitForStreamingRunning();
-
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
 
         ObjectId objId = new ObjectId();
         Document obj = new Document()
@@ -1201,7 +1181,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(FLATTEN_STRUCT, "true");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1253,7 +1232,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> props = new HashMap<>();
         props.put(FLATTEN_STRUCT, "true");
         props.put(DELIMITER, "-");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1305,7 +1283,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> props = new HashMap<>();
         props.put(FLATTEN_STRUCT, "true");
         props.put(DELIMITER, "-");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1374,13 +1351,12 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
         assertThat(transformed.headers()).hasSize(1);
-        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Operation.CREATE.code());
     }
 
     @Test
@@ -1390,13 +1366,12 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op,id");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
         assertThat(transformed.headers()).hasSize(2);
-        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, "__op")).isEqualTo(Operation.CREATE.code());
         assertThat(getSourceRecordHeaderByKey(transformed, "__id")).isNull();
     }
 
@@ -1408,13 +1383,12 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "op,source.rs,source.collection");
         props.put(ADD_HEADERS_PREFIX, "prefix.");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
         assertThat(transformed.headers()).hasSize(3);
-        assertThat(getSourceRecordHeaderByKey(transformed, "prefix.op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, "prefix.op")).isEqualTo(Operation.CREATE.code());
         assertThat(getSourceRecordHeaderByKey(transformed, "prefix.source_rs")).isEqualTo("rs0");
         assertThat(getSourceRecordHeaderByKey(transformed, "prefix.source_collection")).isEqualTo(getCollectionName());
     }
@@ -1426,12 +1400,11 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.CREATE.code());
     }
 
     @Test
@@ -1446,15 +1419,14 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         props.put(ADD_FIELDS_PREFIX, fieldPrefix);
         props.put(ADD_HEADERS, "op:OPERATION");
         props.put(ADD_HEADERS_PREFIX, headerPrefix);
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
 
-        assertThat(((Struct) transformed.value()).get(fieldPrefix + "OP")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get(fieldPrefix + "OP")).isEqualTo(Operation.CREATE.code());
         assertThat(transformed.headers()).hasSize(1);
-        assertThat(getSourceRecordHeaderByKey(transformed, headerPrefix + "OPERATION")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(getSourceRecordHeaderByKey(transformed, headerPrefix + "OPERATION")).isEqualTo(Operation.CREATE.code());
     }
 
     @Test
@@ -1465,12 +1437,11 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op , ts_ms");
         props.put(ADD_FIELDS_PREFIX, "prefix.");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("prefix.op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("prefix.op")).isEqualTo(Operation.CREATE.code());
         assertThat(((Struct) transformed.value()).get("prefix.ts_ms")).isNotNull();
     }
 
@@ -1481,12 +1452,11 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op,id");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.CREATE.code());
         assertThat(((Struct) transformed.value()).get("__id")).isNull();
     }
 
@@ -1497,12 +1467,11 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "op,source.rs,source.collection");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecord createRecord = createCreateRecord();
         final SourceRecord transformed = transformation.apply(createRecord);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.CREATE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.CREATE.code());
         assertThat(((Struct) transformed.value()).get("__source_rs")).isEqualTo("rs0");
         assertThat(((Struct) transformed.value()).get("__source_collection")).isEqualTo(getCollectionName());
     }
@@ -1513,14 +1482,14 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op");
         transformation.configure(props);
 
         final SourceRecord deleteRecord = createDeleteRecordWithTombstone().allRecordsInOrder().get(0);
         final SourceRecord transformed = transformation.apply(deleteRecord);
         assertThat(((Struct) transformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
     }
 
     @Test
@@ -1529,14 +1498,14 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op,ts_ms");
         transformation.configure(props);
 
         final SourceRecord deleteRecord = createDeleteRecordWithTombstone().allRecordsInOrder().get(0);
         final SourceRecord transformed = transformation.apply(deleteRecord);
         assertThat(((Struct) transformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
         assertThat(((Struct) transformed.value()).get("__ts_ms")).isNotNull();
     }
 
@@ -1546,14 +1515,14 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op,source.rs,source.collection");
         transformation.configure(props);
 
         final SourceRecord deleteRecord = createDeleteRecordWithTombstone().allRecordsInOrder().get(0);
         final SourceRecord transformed = transformation.apply(deleteRecord);
         assertThat(((Struct) transformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
+        assertThat(((Struct) transformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
         assertThat(((Struct) transformed.value()).get("__source_rs")).isEqualTo("rs0");
         assertThat(((Struct) transformed.value()).get("__source_collection")).isEqualTo(getCollectionName());
     }
@@ -1564,8 +1533,9 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
+        props.put(HANDLE_DELETES, "rewrite");
         props.put(ADD_FIELDS, "op,ts_ms");
+        props.put(DROP_TOMBSTONE, "false");
         transformation.configure(props);
 
         final SourceRecords records = createDeleteRecordWithTombstone();
@@ -1573,7 +1543,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final SourceRecord deleteRecord = records.allRecordsInOrder().get(0);
         final SourceRecord deleteTransformed = transformation.apply(deleteRecord);
         assertThat(((Struct) deleteTransformed.value()).get("__deleted")).isEqualTo(true);
-        assertThat(((Struct) deleteTransformed.value()).get("__op")).isEqualTo(Envelope.Operation.DELETE.code());
+        assertThat(((Struct) deleteTransformed.value()).get("__op")).isEqualTo(Operation.DELETE.code());
         assertThat(((Struct) deleteTransformed.value()).get("__ts_ms")).isNotNull();
 
         final SourceRecord tombstoneRecord = records.allRecordsInOrder().get(1);
@@ -1587,7 +1557,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> transformationConfig = new HashMap<>();
         transformationConfig.put("array.encoding", "array");
         transformationConfig.put("field.name.adjustment.mode", "avro");
-        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -1641,7 +1610,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
 
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_FIELDS, "updateDescription.updatedFields");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         // Perform transformation
@@ -1685,7 +1653,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> props = new HashMap<>();
         props.put(ARRAY_ENCODING, "array");
         props.put(ADD_FIELDS, "op,source.ts_ms");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         final SourceRecords records = createCreateRecordFromJson("dbz-2316.json");
@@ -1699,7 +1666,8 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     public void testMatrixType() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
         transformationConfig.put("array.encoding", "array");
-        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
+        transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
+        transformationConfig.put(HANDLE_DELETES, "none");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -1794,7 +1762,8 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     public void testMatrixArrayAsDocumentType() throws InterruptedException, IOException {
         final Map<String, String> transformationConfig = new HashMap<>();
         transformationConfig.put("array.encoding", "document");
-        transformationConfig.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
+        transformationConfig.put(CONFIG_DROP_TOMBSTONES, "false");
+        transformationConfig.put(HANDLE_DELETES, "none");
         transformation.configure(transformationConfig);
 
         // Test insert
@@ -1859,8 +1828,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
     public void shouldSupportNestedArrays() throws InterruptedException {
         waitForStreamingRunning();
 
-        transformation.configure(Collect.hashMapOf(HANDLE_TOMBSTONE_DELETES, "tombstone"));
-
         // Test insert
         try (var client = connect()) {
             client.getDatabase(DB_NAME).getCollection(this.getCollectionName())
@@ -1922,7 +1889,6 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         final Map<String, String> props = new HashMap<>();
         props.put(ADD_HEADERS, "updateDescription.updatedFields,nonexistentField,version");
         props.put(ADD_HEADERS_PREFIX, "prefix.");
-        props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
@@ -1977,7 +1943,7 @@ public class NewExtractNewDocumentStateTestIT extends AbstractExtractNewDocument
         waitForStreamingRunning();
 
         final Map<String, String> props = new HashMap<>();
-        props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
+        props.put(HANDLE_DELETES, "rewrite");
         transformation.configure(props);
 
         ObjectId objId = new ObjectId();
