+++ debezium-core/src/test/java/io/debezium/transforms/ExtractNewRecordStateTest.java
@@ -20,15 +20,13 @@ import org.junit.Test;
 import io.debezium.data.Envelope;
 import io.debezium.doc.FixFor;
 
-/**
- * @author Jiri Pechanec
- */
 public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
     @Test
     public void testTombstoneDroppedByDefault() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             transform.configure(props);
 
             final SourceRecord tombstone = new SourceRecord(new HashMap<>(), new HashMap<>(), "dummy", null, null);
@@ -40,7 +38,8 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testTombstoneDroppedConfigured() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(DROP_TOMBSTONES, "true");
+            // props.put(DROP_TOMBSTONES, "true");
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             transform.configure(props);
 
             final SourceRecord tombstone = new SourceRecord(new HashMap<>(), new HashMap<>(), "dummy", null, null);
@@ -52,7 +51,22 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testTombstoneForwardConfigured() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(DROP_TOMBSTONES, "false");
+            // props.put(DROP_TOMBSTONES, "false");
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
+            transform.configure(props);
+
+            final SourceRecord tombstone = new SourceRecord(new HashMap<>(), new HashMap<>(), "dummy", null, null);
+            // assertThat(transform.apply(tombstone)).isEqualTo(tombstone);
+            // convert delete event to tombstone record and skip the following tombstone if existed
+            assertThat(transform.apply(tombstone)).isNull();
+        }
+    }
+
+    @Test
+    public void testTombstoneRewriteWithTombstone() {
+        try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
+            final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
             transform.configure(props);
 
             final SourceRecord tombstone = new SourceRecord(new HashMap<>(), new HashMap<>(), "dummy", null, null);
@@ -64,6 +78,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testTruncateDroppedByDefault() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             transform.configure(props);
 
             final SourceRecord truncate = createTruncateRecord();
@@ -75,6 +90,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testDeleteDroppedByDefault() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             transform.configure(props);
 
             final SourceRecord deleteRecord = createDeleteRecord();
@@ -86,7 +102,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testHandleDeleteDrop() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "drop");
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             transform.configure(props);
 
             final SourceRecord deleteRecord = createDeleteRecord();
@@ -98,7 +114,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testHandleDeleteNone() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "none");
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             transform.configure(props);
 
             final SourceRecord deleteRecord = createDeleteRecord();
@@ -111,7 +127,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testHandleDeleteRewrite() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
             transform.configure(props);
 
             final SourceRecord deleteRecord = createDeleteRecord();
@@ -124,7 +140,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testHandleCreateRewrite() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
             props.put(ADD_HEADERS, "op");
             transform.configure(props);
 
@@ -141,6 +157,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testUnwrapCreateRecord() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             transform.configure(props);
 
             final SourceRecord createRecord = createCreateRecord();
@@ -155,6 +172,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testUnwrapCreateRecordWithOptionalDefaultValue() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             transform.configure(props);
 
             final SourceRecord createRecord = createCreateRecordWithOptionalNull();
@@ -168,6 +186,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testIgnoreUnknownRecord() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             transform.configure(props);
 
             final SourceRecord unknownRecord = createUnknownRecord();
@@ -183,6 +202,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testUnwrapPropagatesRecordHeaders() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             transform.configure(props);
 
             final SourceRecord createRecord = createCreateRecord();
@@ -203,6 +223,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddField() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_FIELDS, "op");
             transform.configure(props);
 
@@ -225,6 +246,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddTimestamp() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props1 = new HashMap<>();
+            props1.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props1.put(ADD_FIELDS, "ts_ms");
             transform.configure(props1);
 
@@ -247,6 +269,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFields() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_FIELDS, "op , lsn,id");
             props.put(ADD_FIELDS_PREFIX, "prefix.");
             transform.configure(props);
@@ -264,6 +287,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testNewFieldAndHeaderMapping() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             String fieldPrefix = "";
             String headerPrefix = "prefix.";
             props.put(ADD_FIELDS, "op:OP, lsn:LSN, id:ID, source.lsn:source_lsn, transaction.total_order:TOTAL_ORDER");
@@ -297,6 +321,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFieldsForMissingOptionalField() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_FIELDS, "op,lsn,id");
             transform.configure(props);
 
@@ -313,6 +338,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFieldsSpecifyStruct() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_FIELDS, "op,source.lsn,transaction.id,transaction.total_order");
             transform.configure(props);
 
@@ -330,6 +356,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddHeader() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_HEADERS, "op");
             transform.configure(props);
 
@@ -346,6 +373,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddHeaders() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_HEADERS, "op , lsn,id");
             transform.configure(props);
 
@@ -366,6 +394,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddHeadersForMissingOptionalField() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_HEADERS, "op,lsn,id");
             transform.configure(props);
 
@@ -386,6 +415,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddHeadersSpecifyStruct() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_HEADERS, "op,source.lsn,transaction.id,transaction.total_order");
             props.put(ADD_HEADERS_PREFIX, "prefix.");
             transform.configure(props);
@@ -408,6 +438,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddTopicRoutingField() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ROUTE_BY_FIELD, "name");
             transform.configure(props);
 
@@ -421,6 +452,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testUpdateTopicRoutingField() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ROUTE_BY_FIELD, "name");
             transform.configure(props);
 
@@ -434,8 +466,8 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testDeleteTopicRoutingField() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ROUTE_BY_FIELD, "name");
-            props.put(HANDLE_DELETES, "none");
 
             transform.configure(props);
 
@@ -449,9 +481,8 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddHeadersHandleDeleteRewriteAndTombstone() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
             props.put(ADD_HEADERS, "op,source.lsn");
-            props.put(DROP_TOMBSTONES, "false");
             transform.configure(props);
 
             final SourceRecord deleteRecord = createDeleteRecord();
@@ -473,7 +504,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFieldHandleDeleteRewrite() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
             props.put(ADD_FIELDS, "op");
             transform.configure(props);
 
@@ -489,7 +520,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFieldsHandleDeleteRewrite() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
             props.put(ADD_FIELDS, "op,lsn");
             transform.configure(props);
 
@@ -506,9 +537,8 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFieldsHandleDeleteRewriteAndTombstone() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite-with-tombstone");
             props.put(ADD_FIELDS, "op,lsn");
-            props.put(DROP_TOMBSTONES, "false");
             transform.configure(props);
 
             final SourceRecord deleteRecord = createDeleteRecord();
@@ -527,7 +557,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testAddFieldsSpecifyStructHandleDeleteRewrite() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
-            props.put(HANDLE_DELETES, "rewrite");
+            props.put(HANDLE_TOMBSTONE_DELETES, "rewrite");
             props.put(ADD_FIELDS, "op,source.lsn");
             transform.configure(props);
 
@@ -544,6 +574,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
     public void testSchemaChangeEventWithOperationHeader() {
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             props.put(ADD_HEADERS, "op");
             transform.configure(props);
 
@@ -563,6 +594,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             props.put(DROP_FIELDS_HEADER_NAME, "drop-fields");
             transform.configure(props);
 
@@ -630,6 +662,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             props.put(DROP_FIELDS_HEADER_NAME, "drop-fields");
             props.put(DROP_FIELDS_KEEP_SCHEMA_COMPATIBLE, "false");
             transform.configure(props);
@@ -702,6 +735,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             props.put(DROP_FIELDS_HEADER_NAME, "drop-fields");
             props.put(DROP_FIELDS_FROM_KEY, "true");
             transform.configure(props);
@@ -770,6 +804,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "drop");
             props.put(DROP_FIELDS_HEADER_NAME, "drop-fields");
             props.put(DROP_FIELDS_KEEP_SCHEMA_COMPATIBLE, "false");
             props.put(DROP_FIELDS_FROM_KEY, "true");
@@ -841,6 +876,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             String fieldPrefix = "";
             props.put(ADD_FIELDS, "notExisting,op:OP, lsn:LSN, id:ID, source.lsn:source_lsn, transaction.total_order:TOTAL_ORDER, changes:META_SRC_CHANGED");
             props.put(ADD_FIELDS_PREFIX, fieldPrefix);
@@ -865,6 +901,7 @@ public class ExtractNewRecordStateTest extends AbstractExtractStateTest {
 
         try (ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
+            props.put(HANDLE_TOMBSTONE_DELETES, "tombstone");
             String fieldPrefix = "";
             props.put(ADD_FIELDS, "op:OP, lsn:LSN, id:ID, source.lsn:source_lsn, transaction.total_order:TOTAL_ORDER, notExist:META_SRC_CHANGED");
             props.put(ADD_FIELDS_PREFIX, fieldPrefix);
